Added a takeMeasurement method. Un-hardcoded sensor port variable. Tweaked variables.
Added a takeMeasurement method, which can be placed in a periodic method to continuously update the median filter and the IR sensor's voltage outputs.

IR sensor port is no longer hard coded. Channel is passed as an argument into the class' constructor.

Tweaked variables, variable names, and comments, for accuracy.